[NSX] Refactor API wrapper operations
